corex(misc): porting guide + .github CI + NOTICE + io/geometry/pybind tweaks
docs/porting-build-run-guide.md: end-to-end build & run guide for
the Iluvatar Corex 4.4 port. Covers the offline vcpkg layout, the
UIPC_MUDA_USE_COREX / UIPC_COREX_USE_ASYNC_MEMORY CMake options,
the cudaMallocAsync availability story, slope / wrecking_ball
benchmark numbers, and a checklist of corex backend issues
uncovered by apps/tests/.

.github/workflows/: CMake / xmake / docs / python-wheels CI
configurations carried over from the working copy.

NOTICE: upstream attribution for GPU IPC and contributors.

Smaller follow-ups uncovered while running apps/tests on corex:
  - python/setup.py + scripts/stubgen.py: tolerate the corex
    backend layout when generating .pyi stubs.
  - src/io/urdf_io.cpp + src/io/simplicial_complex_io.cpp:
    float32-safe attribute reads.
  - src/geometry/intersection.cpp +
    src/geometry/affine_body/affine_body_from_rigid_body.cpp:
    avoid double/float Eigen mixing that the corex clang-cuda
    wrapper rejects.
  - src/core/diff_sim/sparse_coo_view.cpp: small accessor signature
    change required by the new sparse_coo_view.h.
  - src/pybind/pyuipc/geometry/affine_body.cpp: matching pybind
    binding for the geo_ids attribute exposed by the joint
    constitutions.

Made-with: Cursor

+++ src/geometry/intersection.cpp
@@ -43,7 +43,8 @@ bool tri_edge_intersect(const Vector3& T0,
         if(point_in_triangle(E0, T0, T1, T2) || point_in_triangle(E1, T0, T1, T2))
             return true;
 
-        Float u, v;
+        // libigl segment_segment_intersect only exposes double& outputs
+        double u, v;
         // Check if edge and any triangle edge overlap (as segments)
         if(igl::segment_segment_intersect(E0, E1 - E0, T0, T1 - T0, u, v)
            || igl::segment_segment_intersect(E0, E1 - E0, T1, T2 - T1, u, v)
